+++ packages/client-auth/RELEASES.md
@@ -259,3 +259,13 @@ Based on:
 - [typescript v1.0.0-alpha.25] packages/client-auth
 ### Releases
 - [NPM v1.0.0-alpha.25] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.25 - packages/client-auth
+
+## 2024-08-13 00:09:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.363.1 (2.396.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.26] packages/client-auth
+### Releases
+- [NPM v1.0.0-alpha.26] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.26 - packages/client-auth